pyproject.toml

[tool.poetry]
name = "django-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
django-filter = "^23.5"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
redis = "^5.2.1"
django-celery-beat = "^2.8.0"
eventlet = "^0.39.1"
celery = "^5.5.2"
dotenv = "^0.9.9"
coreapi = "2.3.3"
setuptools = "^68.0.0"
stripe = "^12.1.0"
forex-python = "^1.8"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.2"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
